Refactored format selection, making it automatic
Colors are now sorted out by looking at the given image, and determining
how many unique colors were used.  If only 2 colors were used, then the
tool will assign the "darker" color a value of 0, and the "lighter"
color a value of 1, generate a coe file with radix 2, and then generate
a table for the user to perform lookups on the colors used.

Radix 16 and 10 are also supported, and the utility will simply bin the
image according to the number of unique colors.
1-2           : Radix 2
3-16          : Radix 16
17-4294967296 : Radix 10

If you happen to exceed 4294967296, then I don't know what you're trying
to do on an FPGA anymore.  Create an issue on Github, we can debate
about it for a couple hours, and then we can go to Xilinx and tell them
we exceeded the limitations for a coe file. We can ride through the
storms of hopeless inaction and hyped up weather reports only to find
that either they won't do anything and the storm never comes, or there's
some alternative to coe files that we should be using for one reason or
another and a hurricane knocks us flat on our feet while we were staring
out the window enjoying a nice hot cup of coffee in the morning, trying
to fit the world's most detailed cat picture on a FPGA which sounded
like an amazing idea at the time.